analytics-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
